Super updejt
Logowanie ->
Sprawdza błąd czy dane zostały poprawnie wprowadzone jeżeli zostały źle wprowadzone zwraca error(wiadomość pod logowaniem ze dane zostały zle wprowadzone), jezeli dobrze to przechodzi do home.

Indywidualne dzieci (child) -> Nie było poprawne wyswietlane dane/jakeis zmieniany css był, naprawiłem żeby sie poprawnie zwracało.

Dodanie potwierdzenie usuniecia grupy/rodziców/dzieci w zakladace rodzice + grupy + dzieci (wyskakuje czy chcesz usunac dana wartosc i trzeba potwierdzic zeby zatwierdzic zmiany)

W wiadomosci status - ustawienie żeby zwracało przeczytane/nie przeczytane + kolor wiersza na czerwony/zielony zaleznie od statusu wiadomosci zamiast tego false/true co było.

wysyłanie wiadomosci - dostajemy potwierdzenie czy wiadomosc została wysylana lub nie zaleznie od tego czy zostanie wyslany czy tez wyskoczy error (problem taki ze nie wyskakuje nigdy error bo zawsze sie wysyla do bazy danych nawet jezlei wyslemy puste etc. Wiec teraz zawsze bedzie zwracało ze została wyslana,
jak bedzie zmienione na backu zeby nie akceptowało chociaz pustych wartosci czy cos to juz bedzie git).

z strony głównej usunieta mozliwosc edycji częsci danych, został tylko adres i telefon.  + dodany początek dodania zdjecia profilowego dla uzytkownika, chciałem przez imgura wrzucic tymczasowo ale za duzo srania sie + nie działało koniec koncow

Dodałem możliwość wyszukiwania grup po nazwie + w bazy wiedzy tez dodałem, tak testowo bo mieliśmy miec jakas opcje przeszukiwania elementów probowalem knowledge przerobic na jakas galerie itd. nie za bardzo wychodziło wiec tylko zedytowałem te wsywietlanie informacji + edcyje okienko jakies tam cssy.

Zablokowałem wyświetalnie usuwania plików/wszystkich plików/Dodawania Plików/edycji dla wszystkich poza adminem

Do wiadomości dodałem to wyszukiwanie po imieniu i nazwisku i naprawiłem bo idk czemu nie łapało jak sie wpisywało  x - ilość uzytkownikow, w jakiejs wersji musialem złą wprowadzic czy cos, nie wiem. Tylko sie zaciąłem na tym żeby wyświetlało kolejnych uzytkownikow po dodaniu przecinka, pewnie to jest
jakas drobna zmiana na która mam zbyt wypalony łeb teraz wiec to zostawiłem bo mi psycha od tego siadała.

Menu probowalem troche edytować żeby były prerenderowane wartosci aby uniknac tego blikania ale nic nie dało + mozna dodac niby jakis element ładowania w sensie ze zamiast instant sie ładować strona bedzie jakiś buffer 1/2 sec miedzy stronami co sprawi ze sie zaladuja elementy i pozniej zostana wyswietlone,
niby jakies obejscie na te migajace elementy (nie wiem jak sie to robi jeszcze, tak tylko czytalem sobie).

+++ frontend/wirtualneprzedszkole/src/pages/Children/ChildrenComponent.js
@@ -2,6 +2,7 @@ import React from "react";
 import ChildrenService from "./ChildrenService";
 import "../User/Table.scss";
 import { useNavigate } from "react-router-dom";
+import UserService from "../User/UserService";
 
 const Navi = (props) => {
   const navigate = useNavigate();
@@ -28,13 +29,20 @@ class ChildrenComponent extends React.Component {
     console.log(this.state);
   }
 
+
   deleteChild(id) {
-    ChildrenService.deleteChild(id).then((response) => {
-      this.setState({
-        children: this.state.children.filter((child) => child.id !== id),
+    let childName = this.state.children.find(child => child.id === id).name;
+    let childlastName = this.state.children.find(child => child.id === id).lastName;
+    if(window.confirm("Czy na pewno chcesz usunąć użytkownika: "  + childName +" " + childlastName + " "+ " ?")) {
+      ChildrenService.deleteChild(id).then((response) => {
+        this.setState({
+          children: this.state.children.filter((child) => child.id !== id),
+        });
       });
-    });
+    }
   }
+
+
 
   componentDidMount() {
     ChildrenService.getChildren().then((response) => {
