+++ frontend/wirtualneprzedszkole/src/pages/User/UserComponent.js
@@ -6,12 +6,12 @@ import { useNavigate } from "react-router-dom";
 const Navi = (props) => {
   const navigate = useNavigate();
   return (
-    <button
-      onClick={() => navigate("/user/" + props.value, { replace: true })}
-      className="btn btn-info"
-    >
-      Zobacz
-    </button>
+      <button
+          onClick={() => navigate("/user/" + props.value, { replace: true })}
+          className="btn btn-info"
+      >
+        Zobacz
+      </button>
   );
 };
 
@@ -29,11 +29,16 @@ class UserComponent extends React.Component {
   }
 
   deleteUser(id) {
-    UserService.deleteUser(id).then((response) => {
-      this.setState({
-        users: this.state.users.filter((user) => user.id !== id),
+    let userName = this.state.users.find(user => user.id === id).name;
+    let userlastName = this.state.users.find(user => user.id === id).lastName;
+    let emaail = this.state.users.find(user => user.id === id).email;
+    if(window.confirm("Czy na pewno chcesz usunąć użytkownika: "  + userName +" " + userlastName + " "+ "("+emaail+")" + " ?")) {
+      UserService.deleteUser(id).then((response) => {
+        this.setState({
+          users: this.state.users.filter((user) => user.id !== id),
+        });
       });
-    });
+    }
   }
 
   componentDidMount() {
@@ -45,9 +50,9 @@ class UserComponent extends React.Component {
 
   render() {
     return (
-      <div className="scrollable-div">
-        <table className="content-table">
-          <thead>
+        <div className="scrollable-div">
+          <table className="content-table">
+            <thead>
             <tr className="table-head">
               <td>Id</td>
               <td>Imię</td>
@@ -55,29 +60,29 @@ class UserComponent extends React.Component {
               <td>Email</td>
               <td>Akcje</td>
             </tr>
-          </thead>
-          <tbody className="body table-body">
+            </thead>
+            <tbody className="body table-body">
             {this.state.users.map((user) => (
-              <tr key={user.id}>
-                <td>{user.id}</td>
-                <td>{user.name}</td>
-                <td>{user.lastName}</td>
-                <td>{user.email}</td>
-                <td className="foobar">
-                  <Navi value={user.id} />
-                  {/* <button onClick={() => this.props.navigation.navigate("/home//")} className='btn btn-info'>Zobacz</button> */}
-                  <button
-                    onClick={() => this.deleteUser(user.id)}
-                    className="btn btn-danger"
-                  >
-                    Usuń
-                  </button>
-                </td>
-              </tr>
+                <tr key={user.id}>
+                  <td>{user.id}</td>
+                  <td>{user.name}</td>
+                  <td>{user.lastName}</td>
+                  <td>{user.email}</td>
+                  <td className="foobar">
+                    <Navi value={user.id} />
+                    {/* <button onClick={() => this.props.navigation.navigate("/home//")} className='btn btn-info'>Zobacz</button> */}
+                    <button
+                        onClick={() => this.deleteUser(user.id)}
+                        className="btn btn-danger"
+                    >
+                      Usuń
+                    </button>
+                  </td>
+                </tr>
             ))}
-          </tbody>
-        </table>
-      </div>
+            </tbody>
+          </table>
+        </div>
     );
   }
 }
